move the directives to lit folder
