+++ packages/dialog/lit.d.ts
@@ -1,0 +1,1 @@
+export * from './src/lit/renderer-directives.js';
