+++ packages/dialog/lit.js
@@ -1,0 +1,1 @@
+export * from './src/lit/renderer-directives.js';
